gpt_helper_bot/requirements.txt
aiogram
python-dotenv
pandas
openai